requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
Django==5.2.6
django-cleanup==9.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.0
google-auth==2.40.3
google-genai==1.35.0
gunicorn==21.2.0
h11==0.16.0
h3==4.3.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
numpy==2.2.6
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pytz==2025.2
qrcode==8.2
requests==2.28.2
rsa==4.9.1
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.26.1
tenacity==9.1.2
timezonefinder==6.5.9
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==1.26.20
websockets==15.0.1
